Fix the following bug:
- When dragging the mouse in either listbox, the *first* entry
clicked on is selected rather than the last (but the last one is
highlighted).

This is done by changing the bindtags so that our binding is executed
after the default binding (which sets the 'active' index to the last
item selected), and using 'active' instead of 'anchor' as the index to
ask for.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@7368 6015fed2-1504-0410-9fe1-9d1591cc4771