Added environment for defining PyVM instructions.  This is only used in the libdis.tex file I'm about to check in.
I'm not sure this is really an optimal solution yet, but it may be the best
alternative.  It avoids describing the instructions as either data items or
functions.

This change was discussed with Guido.  (Guido:  Take a look at the LaTeX
output for this; if this is reasonable I'll go ahead and update the perl
code in myformat.perl to match.)


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@9206 6015fed2-1504-0410-9fe1-9d1591cc4771