Fix a horrible race condition -- various routines were storing the most recently opened URL in self.openedurl of the URLopener instance. This doesn't really work if multiple threads share the same opener instance!
Fix: openedurl was actually simply the type prefix (e.g. "http:")
followed by the rest of the URL; since the rest of the URL is
available and the type is effectively determined by where you are in
the code, I can reconstruct the full URL easily, e.g. "http:" + url.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@9750 6015fed2-1504-0410-9fe1-9d1591cc4771