*Sigh*
Add a -q option for "quiet" operation.  Nothing gets out (not even stderr,
since that's tossed to stdout for capture in the .how file).  It's a good
idea to use -l to get the log in case there are problems.  Maybe this should
be the default and add a --verbose option to get the stdout?  There's way
too much junk output.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@10128 6015fed2-1504-0410-9fe1-9d1591cc4771