Fix two unrelated problems in the same area:
(1) reorder the tests for -Olimit 1500 and -OPT:Olimit=0 so that the
latter test is performed first, and if it works, the former test is
skipped.  This should get rid of the problem that the new SGI
compilers accept both but emit a warning about -Olimit 1500.

(2) The DGUX hack was somehow split in two by the Olimit tests,
probably as the result of a non-context diff.  Moved this back
together again, after the Olimit tests.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@11114 6015fed2-1504-0410-9fe1-9d1591cc4771