All import-related code has moved to macimport.c.
There's also new support for importing code fragments: if a file on
sys.path contains a PYD resource with resourcename equal to the name
of the module to be imported this PYD resource should contain a
(pascal) string with the name of a code fragment to load. This allows
freezing Python programs without access to source or a development
environment.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@11147 6015fed2-1504-0410-9fe1-9d1591cc4771