build_html():  Simplify, so there's only one version of the latex2html 	command line.
l2hoption():  Convenience function, to write out an option to the aux. init.
	file only if set.

Added --iconserver option to set the $ICONSERVER variable in latex2html; this
requires the use of an auxillary init. file since this can't be initialized
using a standard latex2html command-line option.  So the aux. init. file is
used for just about all the special options since it has to be written anyway
when $ICONSERVER needs to be set.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@11310 6015fed2-1504-0410-9fe1-9d1591cc4771