patches from Andrew
NOTE: There is still a bug of some sort in the behavior of zlib.  In
at least one case, inflate returns Z_OK (which is typically
interpreted to mean that more output space is needed) when it has
finished inflating a buffer.  This has been reported as a bug to the
zlib maintainers; we may need to change the Python interface.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@12156 6015fed2-1504-0410-9fe1-9d1591cc4771