Make <rfc> no longer an empty element but a container.  The text currently generated by the LaTeX and LaTeX2HTML processes is generated here as well, making it more flexible in the SGML version.
Reduce the <args> element so that <optional> goes away; just use
square brackets to indicate what's optional.  This makes it easier to
read than the LaTeX, and the processor can do any checking it needs to
in order to make sure it's legit.  Possible shortcoming: DSSSL
processors may need more explicit markup.  Can probably hack around it
for this case, but we'll see.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@12393 6015fed2-1504-0410-9fe1-9d1591cc4771