Jonathan Giddy writes:
Here's a patch to fix the race condition, which wasn't fixed by Rob's
patch.  It holds the gethostbyname lock until the results are copied out,
which means that this lock and the Python global lock are held at the same
time.  This shouldn't be a problem as long as the gethostbyname lock is
always acquired when the global lock is not held.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@12944 6015fed2-1504-0410-9fe1-9d1591cc4771