New offerings by Tim Peters; he writes:
IDLE is now the first Python editor in the Universe not confused by my
doctest.py <wink>.

As threatened, this defines IDLE's is_char_in_string function as a
method of EditorWindow.  You just need to define one similarly in
whatever it is you pass as editwin to AutoIndent; looking at the
EditorWindow.py part of the patch should make this clear.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@13424 6015fed2-1504-0410-9fe1-9d1591cc4771