Patch submitted by Toby Dickenson and approved by Mark Hammond. Toby writes:
winmakemakefile.py tries to allow for spaces in the python install
path, by adding quotes around the appropriate filenames. It doesn't
quite get this correct; sometimes the quotes end up in the middle of
the path.

Microsoft's NMAKE version 6.0 is happy with this (!!!!)  unless there
is also a space in the name. I guess most users of freeze on windows
do not use the same path as the binary distribution.

I've tested the following changes on systems with and without a space
in the path.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@13554 6015fed2-1504-0410-9fe1-9d1591cc4771