Changes inspired by Randall Hooper to allow callbacks when an OptionMenu is modified.  Somewhat rewritten and elaborated by myself.
class _setit: The constructor now takes an optional argument
`callback' and stashes this in a private variable.  If set, the
__call__() method will invoke this callback after the variable's value
has changed.  It will pass the callback the value, followed by any
args passed to __call__().

class OptionMenu: The constructor now takes keyword arguments, the
only one that's legally recognized is `command', which can be set to a
callback.  This callback is invoked when the OptionMenu value is set.
Any other keyword argument throws a TclError.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@14506 6015fed2-1504-0410-9fe1-9d1591cc4771