Charles Waldman writes:
"""
In the course of debugging this I also saw that cPickle is
inconsistent with pickle - if you attempt a pickle.load or pickle.dump
on a closed file, you get a ValueError, whereas the corresponding
cPickle operations give an IOError.  Since cPickle is advertised as
being compatible with pickle, I changed these exceptions to match.
"""


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@15195 6015fed2-1504-0410-9fe1-9d1591cc4771