Charles G Waldman:
Follow a suggestion in an /*XXX*/ comment [in com_add()] to speed up
compilation by using supplemental dictionaries to keep track of names
and constants, eliminating quadratic behavior.  With this patch in
place, the time to import a 5000-line file with lots of constants [at
the global level] is reduced from 20 seconds to under 3 on my system.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@15281 6015fed2-1504-0410-9fe1-9d1591cc4771