Moved LONG_LONG #define from longobject.h to here, since it's needed by the following.
typedef in a portable way the Python name for the C9X uintptr_t type.
This latter is the most portable way to spell an integral type to
which a void* can be cast to and back again without losing
information.  Parallel checkin hacks configure to check if the
platform/compiler supports the C9X name.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@16861 6015fed2-1504-0410-9fe1-9d1591cc4771