Since it looks like the dual license clause may be neither necessary nor sufficient to make Python 2.0 compatible with the GPL, we won't bother with it now.
In other words, we're still where we were weeks ago -- CNRI believes
that its license is GPL-compatible, Stallman says it's not.  I'm
trying to arrange a meeting between their lawyers so they can work it
out.  Whether dual licensing is the solution is open at this point.
If it is the (only!) solution, we'll add that to the BeOpen license
for 2.0 final.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@17254 6015fed2-1504-0410-9fe1-9d1591cc4771