Removed the implib_dir instance variable because it is unused.
Removed get_ext_libname() because it is unused.

Fixed get_libraries() to append an '_d' to the python debug
import library. If MSVC is used, do not add 'pythonxx.lib' to
the list of libraries, because this is handled better
by a pragma in config.h.

This should fix bug #115595, but it needs some more testing.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@17771 6015fed2-1504-0410-9fe1-9d1591cc4771