Minimal fix for the complaints about pickling Unicode objects.  (SF bugs #126161 and 123634).
The solution doesn't use the unicode-escape encoding; that has other
problems (it seems not 100% reversible).  Rather, it transforms the
input Unicode object slightly before encoding it using
raw-unicode-escape, so that the decoding will reconstruct the original
string: backslash and newline characters are translated into their
\uXXXX counterparts.

This is backwards incompatible for strings containing backslashes, but
for some of those strings, the pickling was already broken.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@18466 6015fed2-1504-0410-9fe1-9d1591cc4771