Adapted version of SF Patch #103173 by pyretic: make uu.decode work with spaces in filename.
I changed the module to use string methods instead of the string
module.  Also, instead of stripping the last character of the filename
(assuming this is the linefeed), I strip trailing whitespace (assuming
creating files with trailing whitespace in their name cannot possibly
be a wise idea).

(Note that I believe that /F's "workaround for broken uuencoders" is
no longer needed since the recent fix to binascii.c, but I'll leave it
in since it appears pretty harmless.)


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@18660 6015fed2-1504-0410-9fe1-9d1591cc4771