Use rich comparisons to fulfill an old wish: complex numbers now raise exceptions when compared using <, <=, > or >=.
NOTE: This is a tentative change: this means that cmp() involving
complex numbers will raise an exception when the numbers differ, and
that in turn means that e.g. dictionaries and certain other compounds
(e.g. UserLists) containing complex numbers can't be compared either.
So we'll have to decide whether this is acceptable.  The alpha test
cycle is a good time to keep an eye on this!


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@18838 6015fed2-1504-0410-9fe1-9d1591cc4771