SF patch #408326 by Robin Thomas: slice objects comparable, not hashable
This patch changes the behavior of slice objects in the following
manner:

- Slice objects are now comparable with other slice objects as though
they were logically tuples of (start,stop,step). The tuple is not
created in the comparison function, but the comparison behavior is
logically equivalent.

- Slice objects are not hashable. With the above change to being
comparable, slice objects now cannot be used as keys in dictionaries.

[I've edited the patch for style.  Note that this fixes the problem
that dict[i:j] seemed to work but was meaningless.  --GvR]


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@19973 6015fed2-1504-0410-9fe1-9d1591cc4771