Applying SF patch #412553 by Christopher Lee: fix linuxaudiodev handling of EAGAIN.
This may or may not fix the problem for me (Mandrake 7.2 on a Dell
Optiplex GX110 desktop): I can't hear the output, but it does pass the
test now.  It doesn't fix the problem for Fred (Mandrake 7.2 on a Dell
Inspiron 7500 which has the Maestro sound drivers).  Fred suspects
that it's the kernel version in combination with the driver.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@20196 6015fed2-1504-0410-9fe1-9d1591cc4771