Added module synopsis for the chapter summary.
Added documentation for TestCase.assertRaises().

Added text for "Mapping concepts to classes", and merged it into the
front matter instead of separating it into a new section.

Removed use of "assert" in examples.

Update the descriptions to reflect further changes from discussions on
the pyunit-interest list.

Added documentation for the defaultTestLoader object and the
TestLoader methods.

Added the assert*() names for the TestCase helper methods.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@20345 6015fed2-1504-0410-9fe1-9d1591cc4771