Use new _implicitNameOp() to generate name op code for list comprehensions.
Always emit a SET_LINENO 0 at the beginning of the module.  The
builtin compiler does this, and it's much easier to compare bytecode
generated by the two compilers if they both do.

Move the SET_LINENO inside the FOR_LOOP block for list
comprehensions.  Also for compat. with builtin compiler.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@20353 6015fed2-1504-0410-9fe1-9d1591cc4771