Slight adaptation of Michael Hudson's patch to test PyDict_Next() (with modification of existing dict elements!).
This is part of SF patch #409864: lazy fix for Pings bizarre scoping
crash.

The adaptation I made to Michael's patch was to change the error
handling to avoid masking other errors (moving the specific error
message to inside test_dict_inner()), and to insert a test for
dict==NULL at the start.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@20390 6015fed2-1504-0410-9fe1-9d1591cc4771