Enable nested scopes by default.
Although this is a one-character change, more work needs to be done:
the compiler can get rid of a lot of non-nested-scopes code, the
documentation needs to be updated, the future statement can be
simplified, etc.

But this change enables the nested scope semantics everywhere, and
that's the important part -- we can now test code for compatibility
with nested scopes by default.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@21652 6015fed2-1504-0410-9fe1-9d1591cc4771