Applied SF patch #438424 by Josh Cogliati:
Python's logolike module turtle.py did not display
the turtle except when actually drawing lines.
This patch changes the turtle.py module so that
it displays the turtle at all times when tracing is
on. This is similar to the the way that logo works.
When tracing is off the turtle will not be displayed.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@22307 6015fed2-1504-0410-9fe1-9d1591cc4771