Refactor future feature handling
Replace uses of PyCF_xxx with CO_xxx.

Replace individual feature slots in PyFutureFeatures with single
bitmask ff_features.

When flags must be transfered among the three parts of the interpreter
that care about them -- the pythonrun layer, the compiler, and the
future feature parser -- can simply or (|) the definitions.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@22365 6015fed2-1504-0410-9fe1-9d1591cc4771