Fix SF bug #443600:
Change to get/set/del slice operations so that if the object doesn't
support slicing, *or* if either of the slice arguments is not an int
or long, we construct a slice object and call the get/set/del item
operation instead.  This makes it possible to design classes that
support slice arguments of non-integral types.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@22577 6015fed2-1504-0410-9fe1-9d1591cc4771