Fix calculation of hardest_arg.
The argument properties are ordered from easiest to hardest.  The
harder the arg, the more complicated that code that must be generated
to return it from getChildren() and/or getChildNodes().  The old
calculation routine was bogus, because it always set hardest_arg to
the hardness of the last argument.  Now use max() to always set it to
the hardness of the hardest argument.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@23199 6015fed2-1504-0410-9fe1-9d1591cc4771