Keep track of a type's subclasses (subtypes), in tp_subclasses, which is a list of weak references to types (new-style classes).  Make this accessible to Python as the function __subclasses__ which returns a list of types -- we don't want Python programmers to be able to manipulate the raw list.
In order to make this possible, I also had to add weak reference
support to type objects.

This will eventually be used together with a trap on attribute
assignment for dynamic classes for a major speed-up without losing the
dynamic properties of types: when a __foo__ method is added to a
class, the class and all its subclasses will get an appropriate tp_foo
slot function.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@23579 6015fed2-1504-0410-9fe1-9d1591cc4771