Fix for SF bug #472940: can't getattr() attribute shown by dir()
There really isn't a good reason for instance method objects to have
their own __dict__, __doc__ and __name__ properties that just delegate
the request to the function (callable); the default attribute behavior
already does this.

The test suite had to be fixed because the error changes from
TypeError to AttributeError.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@23872 6015fed2-1504-0410-9fe1-9d1591cc4771