Mods by Tony Lownds (patch 490100, slightly massaged by me) to make Tkinter work with Mac OS X Aqua-Tk, all nicely within ifdefs.
The process is not for the faint of heart, though: you need to download
and install the (alfa) Aqua-Tk, obtain a few needed X11 headers from
somewhere else and then everything builds. To run scripts using Tkinter
you must build with --enable-framework, build Python.app in Mac/OSX
and run your Tkinter scripts with that. Then, about half the tests in
Demo/tkinter work (or at least do something).

Checking this in anyway because it shouldn't break anything, and newer
versions of Aqua-Tk will streamline the process.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@24636 6015fed2-1504-0410-9fe1-9d1591cc4771