Second part of fix for bug [#483982] Python 2.2b2 bdist_wininst crashes.
If no external zip-utility is found, the archive is created by the
zipfile module, which behaves different now than in 2.1: if the
zip-file is created in the root directory if the distribution, it will
contain an (empty) version of itself.

This triggered the above bug - so it's better to create the zip-file
far away in the TMP directory.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@24804 6015fed2-1504-0410-9fe1-9d1591cc4771