Change Windows file.truncate() to (a) restore the original file position, and (b) stop trying to prevent file growth.
Beef up the file.truncate() docs.

Change test_largefile.py to stop assuming that f.truncate() moves the
file pointer to the truncation point, and to verify instead that it leaves
the file position alone.  Remove the test for what happens when a
specified size exceeds the original file size (it's ill-defined, according
to the Single Unix Spec).


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@25604 6015fed2-1504-0410-9fe1-9d1591cc4771