Introduce two new flag bits that can be set in a PyMethodDef method descriptor, as used for the tp_methods slot of a type.  These new flag bits are both optional, and mutually exclusive.  Most methods will not use either.  These flags are used to create special method types which exist in the same namespace as normal methods without having to use tedious construction code to insert the new special method objects in the type's tp_dict after PyType_Ready() has been called.
If METH_CLASS is specified, the method will represent a class method
like that returned by the classmethod() built-in.

If METH_STATIC is specified, the method will represent a static method
like that returned by the staticmethod() built-in.

These flags may not be used in the PyMethodDef table for modules since
these special method types are not meaningful in that case; a
ValueError will be raised if these flags are found in that context.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@25975 6015fed2-1504-0410-9fe1-9d1591cc4771