New PYMALLOC_DEBUG function void _PyMalloc_DebugDumpStats(void). This displays stats about the # of arenas, pools, blocks and bytes, to stderr, both used and reserved but unused.
CAUTION:  Because PYMALLOC_DEBUG is on, the debug malloc routine adds
16 bytes to each request.  This makes each block appear two size classes
higher than it would be if PYMALLOC_DEBUG weren't on.

So far, playing with this confirms the obvious:  there's a lot of activity
in the "small dict" size class, but nothing in the core makes any use of
the 8-byte or 16-byte classes.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@26086 6015fed2-1504-0410-9fe1-9d1591cc4771