Stopped all warnings from the HTML Help Compiler, by generating proper HTML (or, at least, proper in its view).  The TOC file is now identical to what the HTML compiler itself generates, except for whitespace and a glitch identified below.  The pretty-printing done by prechm.py is pretty much destroyed for now; if you need it pretty-printed, just make the Help Compiler save the files (it's got its own idea of pretty- printing anyway).
Glitch:  The title of Ref Man "2.1.6 Blank lines" shows up as a blank
for now.  This is because the relevant entry in ref/index.html contains
nested anchors, and pychm really has no idea what to do with that.  I
hacked it for now to avoid any error messages or worse insanity, and
filed a bug report against the docs.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@26556 6015fed2-1504-0410-9fe1-9d1591cc4771