Fixed a bug in the splitting of lines, and improved the splitting for single byte character sets.  Also fixed a semantic problem with the constructor's default arguments.  Specifically,
__init__(): Change the maxlinelen argument default to None instead of
MAXLINELEN.  The semantics should have been (and now are) that if
maxlinelen is given it is always honored.  If it isn't given, but
header_name is given, then the maximum line length is calculated.  If
neither are given then the default 76 characters is used.

_split(): If the character set is a single byte character set then we
can split the line at the maxlinelen because we know that encoding the
header won't increase its length.  If the charset isn't a single byte
charset then we use the quicker divide-and-conquer line splitting
algorithm as before.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@26837 6015fed2-1504-0410-9fe1-9d1591cc4771