This is a Python 2.1 and 2.2 bugfix candidate: (or how do I "mark" something to be a candidate?)
fixed an old buglet that caused bdb to be unable to
continue in the botframe, after a breakpoint was set.
the key idea is not to set botframe to the bottom level frame,
but its f_back, which actually might be None.
Additional changes: migrated old exception trick to use
sys._getframe(), which exists both in 2.1 and 2.2 .

Note: I believe Mark Hammond needs to look over his code now.
F5 correctly starts up in the debugger, but later on doesn't stop at a given
breakpoint any longer.

kind regards - chris


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@26917 6015fed2-1504-0410-9fe1-9d1591cc4771