Fix non-blocking connect() for Windows.  Refactored the code that retries the connect() call in timeout mode so it can be shared between connect() and connect_ex(), and needs only a single #ifdef.
The test for this was doing funky stuff I don't approve of,
so I removed it in favor of a simpler test.  This allowed me
to implement a simpler, "purer" form of the timeout retry code.
Hopefully that's enough (if you want to be fancy, use non-blocking
mode and decode the errors yourself, like before).


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@27186 6015fed2-1504-0410-9fe1-9d1591cc4771