Cleanup yielding a small speed boost:  before rich comparisons were introduced, list.sort() was rewritten to use only the "< or not <?" distinction.  After rich comparisons were introduced, docompare() was fiddled to translate a Py_LT Boolean result into the old "-1 for <, 0 for ==, 1 for >" flavor of outcome, and the sorting code was left alone.  This left things more obscure than they should be, and turns out it also cost measurable cycles.
So:  The old CMPERROR novelty is gone.  docompare() is renamed to islt(),
and now has the same return conditinos as PyObject_RichCompareBool.  The
SETK macro is renamed to ISLT, and is even weirder than before (don't
complain unless you want to maintain the sort code <wink>).

Overall, this yields a 1-2% speedup in the usual (no explicit function
passed to list.sort()) case when sorting arrays of floats (as sortperf.py
does).  The boost is higher for arrays of ints.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@27693 6015fed2-1504-0410-9fe1-9d1591cc4771