The email package's tests live much better in a subpackage (i.e. email.test), so move the guts of them here from Lib/test.  The latter directory will retain stubs to run the email.test tests using Python's standard regression test.
test_email_torture.py is a torture tester which will not run under
Python's test suite because I don't want to commit megs of data to
that project (it will fail cleanly there).  When run under the mimelib
project it'll stress test the package with megs of message samples
collected from various locations in the wild.

email/test/data is a copy of Lib/test/data.  The fate of the latter is
still undecided.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@27717 6015fed2-1504-0410-9fe1-9d1591cc4771