SF patch #581414: info reader bug
The "Matching vs. Searching" Info node is unreachable from the Info
program (but is fine in Emacs's Info mode). This patch seems to fix
it.  This is the only occurrence where the info reader fails, so
probably it could be addressed in the python docs as a workaround.
Forwarded the report to the info maintainer.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@27829 6015fed2-1504-0410-9fe1-9d1591cc4771