Further SET_LINENO reomval fixes.  See comments in patch #587933.
Use a slightly different strategy to determine when not to call the line
trace function.  This removes the need for the RETURN_NONE opcode, so
that's gone again.  Update docs and comments to match.

Thanks to Neal and Armin!

Also add a test suite.  This should have come with the original patch...


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@28494 6015fed2-1504-0410-9fe1-9d1591cc4771