__version__: Bump to 2.4
Move the imports of Parser and Message inside the
message_from_string() and message_from_file() functions.  This way
just "import email" won't suck in most of the submodules of the
package.

Note: this will break code that relied on "import email" giving you a
bunch of the submodules, but that was never documented and should not
have been relied on.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@28881 6015fed2-1504-0410-9fe1-9d1591cc4771