Commit fix for SF 603831.
Strangely, two out of three patches there seem already committed; but
the essential one (get rid of the assert in object_filenames in
ccompiler.py) was not yet applied.

This makes the build procedure for Twisted work again.

This is *not* a backport candidate despite the fact that identical
code appears to exist in 2.2.2; Twisted builds fine there, so there
must have been a change elsewhere.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@28969 6015fed2-1504-0410-9fe1-9d1591cc4771