Copied the sources from the distutils CVS repository.
Changed the MSVC project file to create the exe in the
lib/distutils/command directory, bdist_wininst.py must still be
changed to use it.

Also changed to use the same zlib as the zlib module - this has the nice
sideeffect that now the buggy 1.1.3 version is no longer used.

Most of the source files now conform to PEP 7, except for the maximum
line length.  Windows api programming in 78 character lines =:(.

README.txt is a new file, but still empty except for placeholders.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@29833 6015fed2-1504-0410-9fe1-9d1591cc4771