Update way a subprocess is launched for Mac OS X.
Another applet mechanism has been developed for Python on Mac OS X and
trying to use the -c "__import__('run').main()" trick is just not working.

macosx_main.py is a new file which should be used as the startup file for
Mac OS X applet bundles. This startup file understands a -p option, which
when seen will start run.main(). When running as an applet, this seems like
the best approach.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@30208 6015fed2-1504-0410-9fe1-9d1591cc4771