M PyShell.py M idle M idle.py M idle.pyw M setup.py
Switch back to installing IDLE as a package.  The IDLE GUI and the
subprocess will both attempt to start up via the package mechanism, but if
IDLE is not yet installed it is possible to run by calling python idle.py
in the IDLE source directory, or to add the source directory to sys.path.

One advantage of doing it this way is IDLE stays off sys.path.

Developed in collaboration with Tony Lownds.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@30252 6015fed2-1504-0410-9fe1-9d1591cc4771