Raise PicklingError when __reduce__() fails, and add memoize() helper function to update the memo.
The first element of the tuple returned by __reduce__() must be a
callable.  If it isn't the Unpickler will raise an error.  Catch this
error in the pickler and raise the error there.

The memoize() helper also has a comment explaining how the memo
works.  So methods can't use memoize() because the write funny codes.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@30885 6015fed2-1504-0410-9fe1-9d1591cc4771