Merge with PyXML 1.28: Wrap a lot of long lines. Clean up a handler for expat.error.
If a lexical handler is set, make sure we call the startDTD() and
endDTD().  If the lexical handler is unset (by setting it to None),
remove the handlers from the underlying pyexpat parser object.
Closes SF bug #485584.

In namespaces mode, make sure we set up the qnames dictionary
correctly for the AttributesNSImpl instance passed to the
start-element-handler.
Closes SF bug #563399.

Support skippedEntity. Fixes #665486.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@30896 6015fed2-1504-0410-9fe1-9d1591cc4771