dis():  Not all opcodes are printable anymore, so print the repr of the opcode character instead (but stripping the quotes).
Added a proto 2 test section for the canonical recursive-tuple case.
Note that since pickle's save_tuple() takes different paths depending on
tuple length now, beefier tests are really needed (but not in pickletools);
the "short tuple" case tried here was actually broken yesterday, and it's
subtle stuff so needs to be tested.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@30995 6015fed2-1504-0410-9fe1-9d1591cc4771