Merge of the folding-reimpl-branch.  Specific changes,
_max_append(): Change the comparison so that the new string is
concatenated if it's less than or equal to the max length.

header_encode(): Allow for maxlinelen == None to mean, don't do any
line splitting.  This is because this module is mostly used by higher
level abstractions (Header.py) which already ensures line lengths.  We
do this in a cheapo way by setting the max_encoding to some insanely
<100k wink> large number.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@31742 6015fed2-1504-0410-9fe1-9d1591cc4771