Make staticmethods and classmethods participate in GC.
If a class was defined inside a function, used a static or class
method, and used super() inside the method body, it would be caught in
an uncollectable cycle.  (Simplified version: The static/class method
object would point to a function object with a closure that referred
to the class.)

Bugfix candidate.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@32164 6015fed2-1504-0410-9fe1-9d1591cc4771