Include a link to the pybsddb web documentation for the modern object oriented DbEnv & Db object BerkeleyDB interface.
Include a note about Sleepycat's BerkeleyDB license with regards to
distributing it within non-opensource applications and include a link
to their website for license details.

Document that 'r' is the default flag for the legacy bsddb btopen(),
hashopen(), and rnopen() functions.  This is apparently different than the
dbm libraries in other languages according to Bug #732951.  Changing
the default from 'r' to 'c' would break backwards compatibility with
legacy bsddb applications; documenting the default should suffice.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@32936 6015fed2-1504-0410-9fe1-9d1591cc4771