list_resize() now has an "exact" option for bypassing the overallocation scheme in situations that likely won't benefit from it.  This further improves memory utilization from Py2.3 which always over-allocates except for PyList_New().
Situations expected to benefit from over-allocation:
    list.insert(), list.pop(), list.append(), and list.extend()

Situations deemed unlikely to benefit:
    list_inplace_repeat, list_ass_slice, list_ass_subscript

The most gray area was for listextend_internal() which only runs
when the argument is a list or a tuple.  This could be viewed as
a one-time fixed length addition or it could be viewed as wrapping
a series of appends.  I left its over-allocation turned on but
could be convinced otherwise.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@35421 6015fed2-1504-0410-9fe1-9d1591cc4771