SF patch #969180, hotshot incorrectly computes elapsed time by Jason Beardsley.
If the seconds are different, we still need to calculate the differences
between milliseconds.

Also, on a Gentoo Linux (2.6.5) dual Athlon MP box with glibc 2.3,
time can go backwards.  This probably happens when the process switches
the CPU it's running on.  Time can also go backwards when running NTP.
If we detect a negative time delta (ie, time went backwards), return
a delta of 0.  This prevents an illegal array access elsewhere.
I think it's safest to *not* update prev_timeofday in this case, so we
return without updating.

Backport candidate.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@36166 6015fed2-1504-0410-9fe1-9d1591cc4771