setup.py now uses the library directories specified in LDFLAGS (``-L`` directories) and the include directories specified in CPPFLAGS (``-I`` directories) for compiling the extension modules.
This has led to the core being compiled with the values in the shell's
CPPFLAGS.  It has also removed the need for special casing to use Fink and
DarwinPorts under darwin since the needed directories can now be specified in
LDFLAGS and CPPFLAGS (e.g., DarwinPorts users can now do
``LDFLAGS=-L/opt/local/lib; CPPFLAGS=-I/opt/local/include; ./configure`` for
everything to work properly).

Parsing the values in the environment variables is done with getopt.  While optparse
would have been a nicer solution it cannot be used because of dependency issues
at execution time; optparse uses gettext which uses struct which will not have
been compiled when the code is imported.  If optparse ever makes its
importation of gettext optional by catching ImportError and setting _() to an
identity function then it can be used.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@37988 6015fed2-1504-0410-9fe1-9d1591cc4771