[ sf.net patch # 1121611 ]
A new hashlib module to replace the md5 and sha modules.  It adds
support for additional secure hashes such as SHA-256 and SHA-512.  The
hashlib module uses OpenSSL for fast platform optimized
implementations of algorithms when available.  The old md5 and sha
modules still exist as wrappers around hashlib to preserve backwards
compatibility.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@39316 6015fed2-1504-0410-9fe1-9d1591cc4771