Backout the last hack and add in this new one. The failure definitely seems timing related.  This change *seems* to work. Since the failure isn't doesn't occur consistently, it's hard to tell.
Running these tests on Solaris in this order:
	test_urllibnet test_operator test_cgi \
	test_isinstance test_future test_ast test_logging

generally caused a failure (about 50% of the time) before the sleep.
I couldn't provoke the failure with the sleep.

This should really be cleaned up by using threading.Events or something
so it is not timing dependent and doesn't hang forever on failure.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@42845 6015fed2-1504-0410-9fe1-9d1591cc4771