Changed comments to make sense now that the LazyList-based examples no longer require any explicit closing to avoid leaking.
That the tee-based examples still do is (I think) still a
mystery.  Part of the mystery is that gc.garbage remains
empty:  if it were the case that some generator in a trash
cycle said it needed finalization, suppressing collection
of that cycle, that generator _would_ show up in gc.garbage.

So this is acting more like, e.g., some tp_traverse slot
isn't visiting all the pointers it should (in which case
the skipped pointer(s) would act like an external root,
silently suppressing collection of everything reachable
from it(them)).


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@45410 6015fed2-1504-0410-9fe1-9d1591cc4771