Fix bug #1000914 (again).
This patches a file that is generated by bgen, however the code is now the
same as a current copy of bgen would generate.  Without this patch most types
in the Carbon.CF module are unusable.

I haven't managed to coax bgen into generating a complete copy of _CFmodule.c 
yet :-(, hence the manual patching.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@46054 6015fed2-1504-0410-9fe1-9d1591cc4771