Bug #1599782: Fix segfault on bsddb.db.DB().type().
The problem is that _DB_get_type() can't be called without the GIL
because it calls a bunch of PyErr_* APIs when an error occurs.
There were no other cases in this file that it was called without the GIL.
Removing the BEGIN/END THREAD around _DB_get_type() made everything work.

Will backport.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@52811 6015fed2-1504-0410-9fe1-9d1591cc4771