Port r54805 from python25-maint branch:
Add code to read from master_fd in the parent, breaking when we get an OSError
(EIO can occur on Linux) or there's no more data to read.  Without this,
test_pty.py can hang on the waitpid() because the child is blocking on the
stdout write.  This will definitely happen on Mac OS X and could potentially
happen on other platforms.  See the comment for details.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@54807 6015fed2-1504-0410-9fe1-9d1591cc4771