Patch # 1140 (my code, approved by Effbot).
Make sure the type of the return value of re.sub(x, y, z) is the type
of y+x (i.e. unicode if either is unicode, str if they are both str)
even if there are no substitutions or if x==z (which triggered various
special cases in join_list()).

Could be backported to 2.5; no need to port to 3.0.


git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@58098 6015fed2-1504-0410-9fe1-9d1591cc4771