Added PyNumber_ToBase and supporting routines _PyInt_Format and _PyLong_Format.  In longobject.c, changed long_format to _PyLong_Format.  In intobject.c, changed uses of PyOS_snprintf to _PyInt_Format instead.
_PyLong_Format is similar to py3k's routine of the same name, except
it has 2 additional parameters: addL and newstyle.  addL was existing
in long_format, and controls adding the trailing "L".  This is
unneeded in py3k.  newstyle is used to control whether octal prepends
"0" (the pre-2.6 style), or "0o" (the 3.0 sytle).

PyNumber_ToBase is needed for PEP 3127 (Integer Literal Support and
Syntax) and PEP 3101 (Advanced String Formatting).

This changeset does not need merging into py3k.



git-svn-id: svn+ssh://pythondev@svn.python.org/python/trunk@60707 6015fed2-1504-0410-9fe1-9d1591cc4771