Bug 1503: Get the test to pass on OSX.  This should make the test more reliable, but I'm not convinced it is the right solution.  We need to determine if this causes the test to hang on any platforms or do other bad things.
Even if it gets the test to pass reliably, it might be that we want
to fix this in socket.  The socket returned from accept() is different
on different platforms (inheriting attributes or not) and we might
want to ensure that the attributes (at least blocking) is the same
across all platforms.


git-svn-id: svn+ssh://svn.python.org/python/trunk@61993 6015fed2-1504-0410-9fe1-9d1591cc4771