Issue 8143: sync unquote in urlparse with urllib; add comment about doing so.
unquote is duplicated in the two files to avoid a circular reference.
(This is fixed in Python3.)  Updates keep getting made to the public unquote
without fixing the urlparse one, however, so this fix syncs the two
and adds a comment to both to make sure changes are applied to both.



git-svn-id: svn+ssh://svn.python.org/python/trunk@81518 6015fed2-1504-0410-9fe1-9d1591cc4771