Add a protocol for determining the mapping of tasks to children
This way if a child unexpectedly dies, the parent knows which tasks
have been lost and can respond accordingly.  (Sorry for the large
commit, but as far as I can tell any subset of this commit would
either just be broken or result in leaking of pid's in some way.)

TODO: deal with the case of a child dying while holding a queue
lock.
TODO: make threadpool work